pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "django-flex-menu"
dynamic = ["version"]
authors = [{name = "Saman Zand Haghighi", email = "samanzandh@gmail.com"}]
license = {text = "BSD"}
description = "A server-side menu for django."
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = ["django>=3.2", "django-mptt>=0.14.0", "django-waffle>=4.0.0"]

[project.urls]
Homepage = "http://github.com/saman-zand-h/django-flex-menu"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["conf*"]

[tool.setuptools.dynamic]
version = {attr = "flex_menu.__version__"}

[tool.mypy]
python_version = "3.8"
exclude = "waffle/tests"
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["django.*"]
ignore_missing_imports = true